Use the correct cfg attribute for debug assertions, and move the ast check code to a new file
